Fixed loadstring
It is now possible to compile Lua code from within another Lua script,
and run it on the same scope instance (previously not possible). This
allows the following native Lua methods to function correctly:
dofile
loadstring
loadfile

This required the addition of a new method to Generator (CompileInline,
which compiles a method which takes no parameters, and can only be run
on the scope it was compiled for); the addition of trace information for
the current LuaScope and ScopeStorage objects; and the addition of the
GetParent() and GetRoot() methods on LuaScope.