Added build script for IronLua
Added a MSBuild script for IronLua. It will compile the solution and
generate Zip, Msi and NuGet packages. The script makes use of tools
from the IronLanguages project. http://github.com/IronLanguages/main
The idea is to slot IronLua in with the other languages in that
project and reuse some of the support infrastructure it provides.